First Draft
This is a two way setup - 1 sender and 1 receiver.
Step 1. Open the sender and wait for the video to be online . Some text will automatically get copied to your clipboard. Share it to the receiver.
Step 2. Once the receiver has those text copied in the clipboard. Open the receiver/answer.html. The receiver's clipboard will get auto updated with some other text.
Step 3. Sender needs to click the 3rd icon from top. And the connection will start.